add slo info
