+++ tecton_client/_internal/utils.py
@@ -4,7 +4,7 @@ from urllib.parse import urlparse
 import httpx
 
 from tecton_client.__about__ import __version__ as tecton_client_version
-from tecton_client._internal.data_types import MetadataOptions, RequestOptions
+from tecton_client._internal.request_utils import MetadataOptions, RequestOptions
 
 
 def get_default_headers(api_key):
